requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
Django==4.1.5
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
mplcursors==0.5.2
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
plotly==5.11.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
requests==2.28.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==8.1.0
urllib3==1.26.13
